Move imageCopier.copyBlobFromStream into a new copy/blob.go
The over-300-line-long function would benefit from splitting up,
and copy/copy.go has grown entirely too long.

As a first step, move it to a new copy/blob.go.

Only moves unmodified code, should not change behavior.
